end of sec 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   container = 'Unread';
   loading = false;
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -30,11 +31,17 @@ export class MessagesComponent implements OnInit {
   }
   pageChanged(event) {
     this.pageNumber = event.page;
-    //this.loadMessages();
+    console.log(event.page);
+    
+    this.loadMessages();
   }
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
+    this.confirmService.confirm('Confirm delete message','this cannot be undone').subscribe(result=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
+        });
+      }
     })
   }
 }
